# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Use API to call the music generation service of Suno.ai and easily integrate it into agents like GPTs.",
    "version": "1.1.1"
  },
...
  "paths": {
...
    "/api/get": {
      "get": {
        "summary": "Get audio information",
        "description": "",
        "tags": ["default"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Pagination number",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ids",
            "description": "Audio IDs, separated by commas. Leave blank to return a list of all music.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["0", "1"],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "suno model name, chirp-v3"
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      }
    }
  }
}